test(#zimic): remove unnecessary test parameters
